00-bootcamp-project/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
paramiko==3.1.0
proto-plus==1.22.2
protobuf==4.22.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
